backend/requirements.txt
aiohttp==3.7.0
akismet==1.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==20.2.0
captcha==0.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography
dateparser==1.0.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
dj-rest-auth==1.1.2
Django==3.2.5
django-allauth==0.43.0
django-bigbluebutton==0.2.5
django-cors-headers==3.5.0
django-countries==6.1.3
django-extensions==3.0.9
django-filter==2.4.0
django-guardian==2.4.0
django-jalali==4.0.0
django-jet==1.0.7
django-phonenumber-field==5.0.0
django-phonenumbers==1.0.1
django-private-storage==2.2.2
django-recaptcha==2.0.6
django-rest-passwordreset==1.1.0
django-storages==1.11.1
django-user-sessions==1.7.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
djangorestframework-simplejwt==4.7.2
drf-writable-nested==0.6.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
geoip2==4.1.0
greenlet==1.1.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
jdatetime==3.6.2
jet-bridge-base==1.0.1
Jinja2==2.11.2
Markdown==3.3.1
MarkupPy==1.14
MarkupSafe==1.1.1
maxminddb==2.0.3
multidict==5.0.0
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==20.8
phonenumbers==8.12.9
Pillow
prompt-toolkit==2.0.9
psycopg2-binary==2.8.5
pycparser==2.20
pydotplus==2.0.2
PyJWT
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.4.1
regex==2021.7.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
social-auth-core==4.1.0
SQLAlchemy==1.4.20
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.2
zipp==3.3.0
django-environ
djangorestframework-jwt



